vexgen/backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
pydantic-settings==2.3.0
motor==3.4.0
python-dotenv==1.0.1
python-dateutil==2.9.0
neo4j==5.20.0
univers==30.11.0
xmltodict==0.13.0
GitPython==3.1.43
toml==0.10.2
regex==2024.5.15
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.8.0
aiohttp==3.9.5